Change CellNAPID from a short int to an int.  This will likely anger the RT.

+++ CCA/Components/MPM/ImpMPM.cc
@@ -701,7 +701,7 @@ void ImpMPM::actuallyInitialize(const ProcessorGroup*,
     const Patch* patch = patches->get(p);
     printTask(patches, patch,cout_doing,"Doing actuallyInitialize");
 
-    CCVariable<short int> cellNAPID;
+    CCVariable<int> cellNAPID;
     new_dw->allocateAndPut(cellNAPID, lb->pCellNAPIDLabel, 0, patch);
     cellNAPID.initialize(0);
